+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Dialog: added new props `disableBodyScroll`, `preventCloseOnClick` and `preventCloseOnEscape`
+
 ## [3.1.4][] - 2023-01-26
 
 ### Changed